+++ vers/changelog.md
@@ -6,9 +6,8 @@ This files contains a version history including all changes relevant for semanti
 
 ## Upcoming Changes
 
-*(none)*
-
-
+* **new Feature**: RawTokenExposalMethod now offers 'subject' and 'permittedScopes'
+* **new Feature**: 'RequiredApiPermissionAttribute' was added.
 
 ## v 2.5.0
 released **2024-04-29**, including: